requirements.txt
PyMuPDF
Pillow==11.0.0
tkinterdnd2==0.4.2
